ci: skip images download in CI make dev

+++ scripts/import_sample_data.sh
@@ -11,10 +11,15 @@ wget -O products.tar.gz https://static.openfoodfacts.org/exports/products.random
 tar -xzvf products.tar.gz -C /mnt/podata/products
 rm products.tar.gz
 
-echo "\033[32m------------------ 2/ Retrieve product images -------------------\033[0m";
-wget -O products.images.tar.gz https://static.openfoodfacts.org/exports/products.random-modulo-100000.images.tar.gz 2>&1
-tar -xzvf products.images.tar.gz -C /opt/product-opener/html/images/products/
-rm products.images.tar.gz
+if [[ -z "${SKIP_SAMPLE_IMAGES}" ]]
+then
+    echo "\033[32m------------------ 2/ Retrieve product images -------------------\033[0m";
+    wget -O products.images.tar.gz https://static.openfoodfacts.org/exports/products.random-modulo-100000.images.tar.gz 2>&1
+    tar -xzvf products.images.tar.gz -C /opt/product-opener/html/images/products/
+    rm products.images.tar.gz
+else
+    echo "\033[32m------------------ SKIPPED product images -------------------\033[0m";
+fi
 
 echo "\033[32m------------------ 3/ Import products -------------------\033[0m";
 perl -I/opt/product-opener/lib /opt/product-opener/scripts/update_all_products_from_dir_in_mongodb.pl
